+++ packages/af-utils-react-virtual-headless/src/singletons/Batch/index.js
@@ -1,0 +1,20 @@
+import call from "/utils/call";
+
+const Queue = new Set();
+
+export default {
+    _level: 0,
+    _start() {
+        this._level++;
+    },
+    _end() {
+        if (--this._level === 0) {
+            /*
+                 calls must not call _startBatch from inside.
+            */
+            Queue.forEach(call);
+            Queue.clear();
+        }
+    },
+    _queue: fn => Queue.add(fn)
+};