+++ packages/af-utils-react-virtual-headless/src/utils/call/index.js
@@ -1,0 +1,3 @@
+const call = fn => fn();
+
+export default call;